+++ src/features/ministry/hooks/useMinistryMonthlyRecord.tsx
@@ -322,12 +322,11 @@ const useMinistryMonthlyRecord = ({
       return '0';
     }
 
-    const [hoursDaily, minutesDaily] = hours_fields.split(':').map(Number);
+    const [hoursField, minutesField] = hours_fields.split(':').map(Number);
+    const [hoursCredit, minutesCredit] = hours_credits.split(':').map(Number);
 
-    const [hoursMonthly, minutesMonthly] = hours_credits.split(':').map(Number);
-
-    let totalHours = hoursDaily + hoursMonthly;
-    const totalMinutes = (minutesDaily || 0) + (minutesMonthly || 0);
+    let totalHours = hoursField + hoursCredit;
+    const totalMinutes = (minutesField || 0) + (minutesCredit || 0);
 
     const minutesRemain = totalMinutes % 60;
     totalHours += (totalMinutes - minutesRemain) / 60;
@@ -336,32 +335,38 @@ const useMinistryMonthlyRecord = ({
   }, [publisher, congReport, hours_fields, hours_credits]);
 
   const minutes_remains = useMemo(() => {
-    const minutes = hours_total.split(':').at(1);
+    const minutes = hours_fields.split(':').map(Number).at(1);
 
     return minutes ? +minutes : 0;
-  }, [hours_total]);
+  }, [hours_fields]);
 
   const hours_credit_enabled = useMemo(() => {
     if (!person) return false;
 
-    if (month < currentReportMonth()) {
+    const hasAssignment = person.person_data.assignments.some(
+      (record) =>
+        record._deleted === false &&
+        record.code === AssignmentCode.MINISTRY_HOURS_CREDIT
+    );
+
+    if (!publisher) return hasAssignment;
+
+    if (!isSelf && !read_only) return hasAssignment;
+
+    if (read_only && month < currentReportMonth()) {
       return hours_credits !== '0:00';
     }
 
-    const assignments = person.person_data.assignments.filter(
-      (record) => record._deleted === false
-    );
-
-    const hasAssignment = assignments.find(
-      (record) => record.code === AssignmentCode.MINISTRY_HOURS_CREDIT
-    );
-
-    if (!isSelf) {
-      return hasAssignment;
-    }
-
     return hoursCreditEnabled ? hasAssignment : false;
-  }, [hours_credits, person, isSelf, hoursCreditEnabled, month]);
+  }, [
+    hours_credits,
+    person,
+    isSelf,
+    hoursCreditEnabled,
+    month,
+    publisher,
+    read_only,
+  ]);
 
   return {
     month_name,