+++ src/features/ministry/report/form_S4/submit_report/useSubmitReport.tsx
@@ -8,7 +8,7 @@ import { SubmitReportProps } from './index.types';
 import { displaySnackNotification } from '@services/states/app';
 import { useAppTranslation, useCurrentUser } from '@hooks/index';
 import { getMessageByCode } from '@services/i18n/translation';
-import { currentMonthServiceYear } from '@utils/date';
+import { addMonths } from '@utils/date';
 import {
   congFieldServiceReportSchema,
   userFieldServiceDailyReportSchema,
@@ -34,6 +34,7 @@ import { decryptData, encryptObject } from '@services/encryption';
 import { CongFieldServiceReportType } from '@definition/cong_field_service_reports';
 import { dbDelegatedFieldServiceReportsSave } from '@services/dexie/delegated_field_service_reports';
 import { handleSaveDailyFieldServiceReport } from '@services/app/user_field_service_reports';
+import { formatDate } from '@services/dateformat';
 
 const useSubmitReport = ({ onClose, month, person_uid }: SubmitReportProps) => {
   const { t } = useAppTranslation();
@@ -110,14 +111,14 @@ const useSubmitReport = ({ onClose, month, person_uid }: SubmitReportProps) => {
   };
 
   const handleTransferMinutes = async () => {
-    const nextMonth = currentMonthServiceYear();
+    const nextMonth = formatDate(addMonths(`${month}/01`, 1), 'yyyy/MM');
 
-    let [year, month] = nextMonth.split('/');
+    let [varYear, varMonth] = nextMonth.split('/');
 
     let valid = false;
 
     do {
-      const newMonth = `${year}/${month}`;
+      const newMonth = `${varYear}/${varMonth}`;
 
       const report = monthlyReports.find(
         (record) => record.report_date === newMonth
@@ -132,23 +133,23 @@ const useSubmitReport = ({ onClose, month, person_uid }: SubmitReportProps) => {
       }
 
       if (!valid) {
-        month = String(+month + 1).padStart(2, '0');
+        varMonth = String(+varMonth + 1).padStart(2, '0');
 
-        if (+month === 13) {
-          month = '01';
-          year = String(+year + 1);
+        if (+varMonth === 13) {
+          varMonth = '01';
+          varYear = String(+varYear + 1);
         }
       }
     } while (!valid);
 
-    const nextReportMonth = `${year}/${month}`;
+    const nextReportMonth = `${varYear}/${varMonth}`;
     await handleNextMonthUpdate(nextReportMonth);
 
     const nextReportDate = `${nextReportMonth}/01`;
     await handleNextDateUpdate(nextReportDate);
   };
 
-  const handleSubmitSelf = async () => {
+  const handleSubmitSelf = async (round: boolean) => {
     let report: CongFieldServiceReportType;
 
     if (congReport) {
@@ -165,9 +166,21 @@ const useSubmitReport = ({ onClose, month, person_uid }: SubmitReportProps) => {
     report.report_data.bible_studies = bible_studies;
     report.report_data.comments = comments;
 
+    let fieldHours = +hours_fields.split(':').at(0);
+
+    if (round) {
+      const fieldMinutes = +hours_fields.split(':').at(1);
+      fieldHours = fieldHours + (fieldMinutes >= 30 ? 1 : 0);
+    }
+
+    let creditHours = +hours_credits.split(':').at(0);
+    const creditMinutes = +hours_credits.split(':').at(1);
+
+    creditHours = creditHours + (creditMinutes >= 30 ? 1 : 0);
+
     report.report_data.hours = {
-      credit: { value: +hours_credits.split(':').at(0), approved: 0 },
-      field_service: +hours_fields.split(':').at(0),
+      credit: { value: creditHours, approved: 0 },
+      field_service: fieldHours,
     };
 
     report.report_data.shared_ministry = shared_ministry;
@@ -177,13 +190,24 @@ const useSubmitReport = ({ onClose, month, person_uid }: SubmitReportProps) => {
     await handleSaveFieldServiceReports(report);
   };
 
-  const handleSubmitPublisher = async () => {
+  const handleSubmitPublisher = async (round: boolean) => {
+    let fieldHours = +hours_fields.split(':').at(0);
+
+    if (round) {
+      const fieldMinutes = +hours_fields.split(':').at(1);
+      fieldHours = fieldHours + (fieldMinutes >= 30 ? 1 : 0);
+    }
+
+    let creditHours = +hours_credits.split(':').at(0);
+    const creditMinutes = +hours_credits.split(':').at(1);
+    creditHours = creditHours + (creditMinutes >= 30 ? 1 : 0);
+
     const report = {
       person_uid: person_uid,
       bible_studies,
       comments: comments,
-      hours: +hours_fields.split(':').at(0),
-      hours_credits: +hours_credits.split(':').at(0),
+      hours: fieldHours,
+      hours_credits: creditHours,
       report_month: month,
       shared_ministry,
       updatedAt: new Date().toISOString(),
@@ -213,18 +237,95 @@ const useSubmitReport = ({ onClose, month, person_uid }: SubmitReportProps) => {
     }
   };
 
-  const handleSubmit = async () => {
+  const handleSubmit = async (round = false) => {
     // check if current role is secretary or group overseer
     if (isSecretary || isGroupOverseer || isGroupAdmin) {
-      await handleSubmitSelf();
+      await handleSubmitSelf(round);
     }
 
     if (!isSecretary && !isGroupOverseer && !isGroupAdmin) {
-      await handleSubmitPublisher();
+      await handleSubmitPublisher(round);
     }
 
     if (isSelf) {
       const report = structuredClone(userReport);
+
+      if (round) {
+        const fieldDaily = report.report_data.hours.field_service.daily
+          .split(':')
+          .map(Number)
+          .reduce((acc, current, index) => {
+            if (index === 0) acc = acc + current * 60;
+            if (index > 0) acc = acc + current;
+            return acc;
+          }, 0);
+
+        const fieldMonthly = report.report_data.hours.field_service.monthly
+          .split(':')
+          .map(Number)
+          .reduce((acc, current, index) => {
+            if (index === 0) acc = acc + current * 60;
+            if (index > 0) acc = acc + current;
+            return acc;
+          }, 0);
+
+        const fieldExcessMinutes = (fieldDaily + fieldMonthly) % 60;
+        const fieldMinutes = 60 - fieldExcessMinutes;
+
+        if (fieldExcessMinutes < 30) {
+          const totalMinutes = fieldMonthly - fieldExcessMinutes;
+          const reportMinute = totalMinutes % 60;
+          const reportHour = (totalMinutes - reportMinute) / 60;
+
+          report.report_data.hours.field_service.monthly = `${reportHour}:${String(reportMinute).padStart(2, '0')}`;
+        }
+
+        if (fieldExcessMinutes >= 30) {
+          const totalMinutes = fieldMonthly + fieldMinutes;
+          const reportMinute = totalMinutes % 60;
+          const reportHour = (totalMinutes - reportMinute) / 60;
+
+          report.report_data.hours.field_service.monthly = `${reportHour}:${String(reportMinute).padStart(2, '0')}`;
+        }
+      }
+
+      const creditDaily = report.report_data.hours.credit.daily
+        .split(':')
+        .map(Number)
+        .reduce((acc, current, index) => {
+          if (index === 0) acc = acc + current * 60;
+          if (index > 0) acc = acc + current;
+          return acc;
+        }, 0);
+
+      const creditMonthly = report.report_data.hours.credit.monthly
+        .split(':')
+        .map(Number)
+        .reduce((acc, current, index) => {
+          if (index === 0) acc = acc + current * 60;
+          if (index > 0) acc = acc + current;
+          return acc;
+        }, 0);
+
+      const creditExcessMinutes = (creditDaily + creditMonthly) % 60;
+      const creditMinutes = 60 - creditExcessMinutes;
+
+      if (creditExcessMinutes < 30) {
+        const totalMinutes = creditMonthly - creditExcessMinutes;
+        const reportMinute = totalMinutes % 60;
+        const reportHour = (totalMinutes - reportMinute) / 60;
+
+        report.report_data.hours.credit.monthly = `${reportHour}:${String(reportMinute).padStart(2, '0')}`;
+      }
+
+      if (creditExcessMinutes >= 30) {
+        const totalMinutes = creditMonthly + creditMinutes;
+        const reportMinute = totalMinutes % 60;
+        const reportHour = (totalMinutes - reportMinute) / 60;
+
+        report.report_data.hours.credit.monthly = `${reportHour}:${String(reportMinute).padStart(2, '0')}`;
+      }
+
       report.report_data.status = 'submitted';
       report.report_data.updatedAt = new Date().toISOString();
 
@@ -278,7 +379,7 @@ const useSubmitReport = ({ onClose, month, person_uid }: SubmitReportProps) => {
     try {
       setIsProcessing(true);
 
-      await handleSubmit();
+      await handleSubmit(true);
 
       displaySnackNotification({
         header: t('tr_done'),