+++ src/features/reports/publisher_records/export_S21/specific_records/active_publishers/index.tsx
@@ -42,7 +42,7 @@ const ActivePublishers = (props: ActivePublishersProps) => {
           variant="main"
           onClick={handleExport}
           disabled={isProcessing}
-          endIcon={isProcessing && <IconLoading />}
+          endIcon={isProcessing && <IconLoading color="var(--white)" />}
         >
           {btnLabel}
         </Button>